refactor: restructur data loader class and fix generic types
