+++ src/Contracts/DataLoaderInterface.php
@@ -7,6 +7,7 @@ namespace AliYavari\PersianFaker\Contracts;
 /**
  * Loads data from a file and retrieves values.
  *
+ * @template TKey as array-key
  * @template TData The expected return type of the loaded data.
  */
 interface DataLoaderInterface
@@ -14,7 +15,7 @@ interface DataLoaderInterface
     /**
      * Retrieves data from the specified file and path.
      *
-     * @return TData
+     * @return array<TKey, TData>
      */
     public function get();
 }
